Migrate 'db' to a multiplatform 'store' module.
This will allow sync to become multiplatform which will eventually be consumed by the Chrome extension.
