--- db/build.gradle
@@ -1,49 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'org.jetbrains.kotlin.android'
-apply plugin: 'kotlin-kapt'
-apply plugin: 'com.squareup.sqldelight'
-
-android {
-  compileSdkVersion buildConfig.compileSdk
-
-  defaultConfig {
-    minSdkVersion buildConfig.minSdk
-
-    testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
-
-    consumerProguardFiles file('src/main/proguard-consumer.pro')
-  }
-
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
-
-  lintOptions {
-    textReport true
-    textOutput 'stdout'
-    lintConfig rootProject.file('lint.xml')
-
-    // We run a full lint analysis as build part in CI, so skip vital checks for assemble task.
-    checkReleaseBuilds false
-  }
-}
-
-dependencies {
-  implementation deps.kotlin.stdlib.jdk
-
-  implementation deps.sqlBrite
-  implementation deps.android.arch.dbFramework
-
-  api deps.rx.java
-
-  implementation deps.timber
-
-  implementation deps.dagger.runtime
-  kapt deps.dagger.compiler
-
-  testImplementation deps.junit
-
-  androidTestImplementation deps.android.test.runner
-  androidTestImplementation deps.truth
-}
