--- db/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.jakewharton.sdksearch.db"/>
