+++ frontend/android/src/main/java/com/jakewharton/sdksearch/ui/ItemDiffer.kt
@@ -1,7 +1,7 @@
 package com.jakewharton.sdksearch.ui
 
 import android.support.v7.util.DiffUtil
-import com.jakewharton.sdksearch.db.Item
+import com.jakewharton.sdksearch.store.Item
 
 internal class ItemDiffer(
   oldQuery: String,
