+++ frontend/android/src/main/java/com/jakewharton/sdksearch/ui/MainActivity.kt
@@ -22,7 +22,7 @@ import com.jakewharton.rxbinding2.widget.textChanges
 import com.jakewharton.sdksearch.R
 import com.jakewharton.sdksearch.api.dac.BaseUrl
 import com.jakewharton.sdksearch.api.dac.DacComponent
-import com.jakewharton.sdksearch.db.DbComponent
+import com.jakewharton.sdksearch.store.DbComponent
 import com.jakewharton.sdksearch.reference.AndroidReference
 import com.jakewharton.sdksearch.reference.ITEM_LIST_URL_PATHS
 import com.jakewharton.sdksearch.reference.PRODUCTION_DAC
