+++ frontend/android/src/main/java/com/jakewharton/sdksearch/ui/OpenDocumentationItemHandler.kt
@@ -6,8 +6,8 @@ import android.content.Intent
 import android.support.customtabs.CustomTabsIntent
 import com.jakewharton.sdksearch.R
 import com.jakewharton.sdksearch.api.dac.BaseUrl
-import com.jakewharton.sdksearch.db.Item
 import com.jakewharton.sdksearch.reference.AndroidReference
+import com.jakewharton.sdksearch.store.Item
 import com.jakewharton.sdksearch.util.toBitmap
 import com.jakewharton.sdksearch.util.toUri
