+++ frontend/android/src/main/java/com/jakewharton/sdksearch/ui/OpenSourceItemHandler.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.support.customtabs.CustomTabsIntent
 import android.widget.Toast
 import com.jakewharton.sdksearch.R
-import com.jakewharton.sdksearch.db.Item
+import com.jakewharton.sdksearch.store.Item
 import com.jakewharton.sdksearch.reference.AndroidReference
 import com.jakewharton.sdksearch.util.toUri
