+++ frontend/android/src/main/java/com/jakewharton/sdksearch/ui/ShareItemHandler.kt
@@ -4,7 +4,7 @@ import android.app.Activity
 import android.support.v4.app.ShareCompat.IntentBuilder
 import com.jakewharton.sdksearch.R
 import com.jakewharton.sdksearch.api.dac.BaseUrl
-import com.jakewharton.sdksearch.db.Item
+import com.jakewharton.sdksearch.store.Item
 import com.jakewharton.sdksearch.util.toUri
 
 internal class ShareItemHandler(
