+++ settings.gradle
@@ -3,7 +3,6 @@ include ':api:dac'
 include ':api:dac:jdk'
 include ':api:dac:js'
 include ':chrome-platform'
-include ':db'
 include ':debug-updater'
 include ':frontend:android'
 include ':frontend:chrome-extension'
@@ -11,4 +10,7 @@ include ':references'
 include ':references:js'
 include ':references:jdk'
 include ':references:validator'
+include ':store'
+include ':store:android'
+include ':store:temp-jdk'
 include ':sync'
