+++ store/android/src/main/proguard-consumer.pro
