+++ store/android/src/main/sqldelight/com/jakewharton/sdksearch/store/Item.sq
