+++ store/build.gradle
@@ -1,0 +1,7 @@
+apply plugin: 'org.jetbrains.kotlin.platform.common'
+
+dependencies {
+  implementation deps.kotlin.stdlib.common
+  testImplementation deps.kotlin.test.common
+  testImplementation deps.kotlin.test.annotations
+}
