+++ store/src/main/kotlin/com/jakewharton/sdksearch/store/ItemStore.kt
@@ -1,0 +1,5 @@
+package com.jakewharton.sdksearch.store
+
+expect interface ItemStore {
+  fun updateListing(listing: String, items: List<Item>)
+}
